server/requirements.txt
fastapi>=0.111,<0.112
uvicorn[standard]>=0.29,<0.32
httpx>=0.27,<0.28
python-dotenv>=1.0,<2.0
pydantic>=2.6,<3.0
